+++ capa/render/result_document.py
@@ -124,22 +124,19 @@ class Metadata(FrozenModel):
         )
 
 
+class CompoundStatementType:
+    AND = "and"
+    OR = "or"
+    NOT = "not"
+    OPTIONAL = "optional"
+
+
 class StatementModel(FrozenModel):
     ...
 
 
-class AndStatement(StatementModel):
-    type = "and"
-    description: Optional[str]
-
-
-class OrStatement(StatementModel):
-    type = "or"
-    description: Optional[str]
-
-
-class NotStatement(StatementModel):
-    type = "not"
+class CompoundStatement(StatementModel):
+    type: str
     description: Optional[str]
 
 
@@ -147,11 +144,6 @@ class SomeStatement(StatementModel):
     type = "some"
     description: Optional[str]
     count: int
-
-
-class OptionalStatement(StatementModel):
-    type = "optional"
-    description: Optional[str]
 
 
 class RangeStatement(StatementModel):
@@ -165,17 +157,15 @@ class RangeStatement(StatementModel):
 class SubscopeStatement(StatementModel):
     type = "subscope"
     description: Optional[str]
-    scope = capa.rules.Scope
+    scope: capa.rules.Scope
 
 
 Statement = Union[
-    OptionalStatement,
-    AndStatement,
-    OrStatement,
-    NotStatement,
-    SomeStatement,
+    # Note! order matters, see #1161
     RangeStatement,
+    SomeStatement,
     SubscopeStatement,
+    CompoundStatement,
 ]
 
 
@@ -185,18 +175,12 @@ class StatementNode(FrozenModel):
 
 
 def statement_from_capa(node: capa.engine.Statement) -> Statement:
-    if isinstance(node, capa.engine.And):
-        return AndStatement(description=node.description)
-
-    elif isinstance(node, capa.engine.Or):
-        return OrStatement(description=node.description)
-
-    elif isinstance(node, capa.engine.Not):
-        return NotStatement(description=node.description)
+    if isinstance(node, (capa.engine.And, capa.engine.Or, capa.engine.Not)):
+        return CompoundStatement(type=node.__class__.__name__.lower(), description=node.description)
 
     elif isinstance(node, capa.engine.Some):
         if node.count == 0:
-            return OptionalStatement(description=node.description)
+            return CompoundStatement(type=CompoundStatementType.OPTIONAL, description=node.description)
 
         else:
             return SomeStatement(